Rename references to kubefed2 with kubefedctl
