+++ docs/concepts.md
@@ -13,7 +13,7 @@
 | Federate             | Federating a set of Kubernetes clusters means, effectively creating a common interface to the pool of these clusters which can be used to deploy Kubernetes applications across those clusters.                                                                                                                     |
 | Federation           | Kubernetes Cluster Federation enables users to federate multiple Kubernetes clusters for resources distribution, service discovery, high availability etc across multiple clusters.                                                                                                                                 |
 | Host Cluster         | A cluster which is used to expose the federation API and run the federation control plane.                                                                                                                                                                                                                          |
-| Cluster Registration | A cluster join the Host Cluster via command `kubefed2 join`.                                                                                                                                                                                                                                                        |
+| Cluster Registration | A cluster join the Host Cluster via command `kubefedctl join`.                                                                                                                                                                                                                                                        |
 | Member Cluster       | A cluster which is registered with the federation API and that federation controllers have authentication credentials for. The Host Cluster can also be a Member Cluster.                                                                                                                                           |
 | ServiceDNSRecord     | A resource that associates one or more Kubernetes Service resources and how to access the Service, with a scheme for constructing Domain Name System (DNS) [resource records](https://www.ietf.org/rfc/rfc1035.txt) for the Service.                                                                                |
 | IngressDNSRecord     | A resource that associates one or more [Kubernetes Ingress](https://kubernetes.io/docs/concepts/services-networking/ingress/) and how to access the Kubernetes Ingress resources, with a scheme for constructing Domain Name System (DNS) [resource records](https://www.ietf.org/rfc/rfc1035.txt) for the Ingress. |
