+++ docs/releasing.md
@@ -12,18 +12,18 @@ Creating a federation v2 release involves the following steps:
 
 1. Locally create an annotated tag in the format `v[0-9]+.[0-9]+.[0-9]+`
    - `git tag -a <tag> -m "Creating release tag <tag>"`
-   - An annotated tag is required to ensure that the `kubefed2` and
+   - An annotated tag is required to ensure that the `kubefedctl` and
      `controller-manager` binaries are versioned correctly.
    - At the time of writing, it is not possible to create an
      annotated tag through the github web interface.
 2. Push the tag to master
    - `git push origin <tag>` (this requires write access to the repo)
-3. Build `kubefed2` for release
-   1. `make kubefed2`
+3. Build `kubefedctl` for release
+   1. `make kubefedctl`
    2. `cd bin` (from repo root)
-   3. `kubefed2 version` (check that the output is as expected)
-   4. `tar cvzf kubefed2.tgz kubefed2`
-   5. `sha256sum kubefed2.tgz > kubefed2.tgz.sha`
+   3. `kubefedctl version` (check that the output is as expected)
+   4. `tar cvzf kubefedctl.tgz kubefedctl`
+   5. `sha256sum kubefedctl.tgz > kubefedctl.tgz.sha`
 4. Package the helm chart for release
    1. Update the default image tag in values.yaml (Change the version to match the release)
    2. Update the chart version in Chart.yaml (Format should be `x.x.x`)
@@ -35,7 +35,7 @@ Creating a federation v2 release involves the following steps:
 4. Create github release
    1. Copy text from old release and replace old tag references
    2. Add a synopsis of the `Unreleased` section of `CHANGELOG.md`
-   3. Add `kubefed2.tgz` and `kubefed2.tgz.sha`
+   3. Add `kubefedctl.tgz` and `kubefedctl.tgz.sha`
    4. Add `federation-v2-<x.x.x>.tgz` and `federation-v2-<x.x.x>.tgz.sha`
 5. Update master
    1. Move the contents of the `Unreleased` section of `CHANGELOG.md` to `v<x.x.x>`
