+++ pkg/controller/sync/resource_test.go
@@ -20,7 +20,7 @@ import (
 	"strings"
 	"testing"
 
-	kfenable "github.com/kubernetes-sigs/federation-v2/pkg/kubefed2/enable"
+	kfenable "github.com/kubernetes-sigs/federation-v2/pkg/kubefedctl/enable"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
