+++ pkg/kubefedctl/enable/schema.go
