+++ pkg/kubefedctl/util/util.go
@@ -38,7 +38,7 @@ type fedConfig struct {
 	pathOptions *clientcmd.PathOptions
 }
 
-// NewFedConfig creates a fedConfig for `kubefed2` commands.
+// NewFedConfig creates a fedConfig for `kubefedctl` commands.
 func NewFedConfig(pathOptions *clientcmd.PathOptions) FedConfig {
 	return &fedConfig{
 		pathOptions: pathOptions,
