+++ pkg/kubefedctl/util/yaml_writer.go
